zpl_generate_service/pyproject.toml
[tool.poetry]
name = "zpl_generate_service"
version = "0.1.0"
description = ""
authors = ["Bernardo Gomes <bgomesdeabreu@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.7"
prettyconf = "^2.2.1"
kafka-python = "^2.0.2"
simple-zpl2 = "^0.3.0"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
